Added integration test for rules_module overrides
