+++ test/integration/chained_dyndest_job_conf.xml
@@ -3,7 +3,7 @@
     <plugins>
         <plugin id="local" type="runner" load="galaxy.jobs.runners.local:LocalJobRunner" workers="2"/>
         <plugin id="dynamic" type="runner">
-            <param id="rules_module">integration.chained_dyndest_rules</param>
+            <param id="rules_module">integration.chained_dyndest_rules.module1</param>
         </plugin>
     </plugins>
 
@@ -19,6 +19,8 @@
         <destination id="dyn_dest2" runner="dynamic">
             <param id="type">python</param>
             <param id="function">dyndest_chain_2</param>
+            <!-- test overriding rules_module at destination level -->
+            <param id="rules_module">integration.chained_dyndest_rules.module2</param>
             <param id="tmp_dir_prefix">from1</param>
         </destination>
     </destinations>
