+++ test/integration/chained_dyndest_rules/module1/rules.py
@@ -1,0 +1,15 @@
+def dyndest_chain_1():
+    # Check whether chaining dynamic job destinations work
+    return "dyn_dest2"
+
+
+def dyndest_chain_2():
+    # Return an invalid destination as this module's function
+    # should never be called
+    return "invalid_destination"
+
+
+def dyndest_chain_3():
+    # Return an invalid destination as this module's function
+    # should never be called
+    return "invalid_destination"
